[fctlog] Lifts copyable args requirement and fixes ostream for unique_ptr
The args are no longer given by value but by reference into the functionlogger.
Thus it is no longer a requirement for them to be copyable.

Furthermore, the function for printing "Containers" was removed and specialized
for vector instead. The previous templating resulted in the method also matching
std::unique_ptr, which then resulted in compile errors. Additionally the previous
implementation also didn't work for any comparison based containers, like std::set.

Furthermore adds a method for printing unique_ptrs.